figma-transformer: extract style declaration helpers
